+++ tests/tt_metal/tt_metal/perf_microbenchmark/CMakeLists.txt
@@ -39,7 +39,7 @@ foreach(TEST_SRC ${PERF_MICROBENCH_TESTS_SRCS})
         target_sources(
             ${TEST_TARGET}
             PRIVATE
-                ${PROJECT_SOURCE_DIR}/tt_metal/impl/experimental/disaggregation/kv_chunk_address_table_protobuf.cpp
+                ${PROJECT_SOURCE_DIR}/tt_metal/impl/internal/disaggregation/kv_chunk_address_table_protobuf.cpp
         )
     endif()
