+++ tt_metal/impl/experimental/disaggregation/kv_chunk_address_table_protobuf.hpp
@@ -2,34 +2,26 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-// Protobuf export/import for KvChunkAddressTable.
-//
-// Consumers must:
-//   1. Generate the proto with GENERATE_PROTO_FILES() and add the .pb.cc to their target
-//   2. Add the protobuf .cpp to their target sources
-//   3. Add appropriate include directories for the generated .pb.h
-//   4. Link protobuf::libprotobuf
-
 #pragma once
 
-#include <string>
+// DEPRECATED: this header has moved to the internal API surface.
+// Include "tt_metal/impl/internal/disaggregation/kv_chunk_address_table_protobuf.hpp"
+// and use tt::tt_metal::internal::disaggregation instead. This shim forwards the
+// serializer free-functions into the old experimental namespace so existing callers
+// keep compiling while they are migrated.
 
-#include "tt_metal/api/tt-metalium/experimental/disaggregation/kv_chunk_address_table.hpp"
+#include "tt_metal/impl/internal/disaggregation/kv_chunk_address_table_protobuf.hpp"
 
 namespace tt::tt_metal::experimental::disaggregation {
 
-// --- Binary wire format ---
+using tt::tt_metal::internal::disaggregation::export_to_protobuf;
+using tt::tt_metal::internal::disaggregation::export_to_protobuf_file;
+using tt::tt_metal::internal::disaggregation::import_from_protobuf;
+using tt::tt_metal::internal::disaggregation::import_from_protobuf_file;
 
-std::string export_to_protobuf(const KvChunkAddressTable& table);
-void export_to_protobuf_file(const KvChunkAddressTable& table, const std::string& path);
-KvChunkAddressTable import_from_protobuf(const std::string& data);
-KvChunkAddressTable import_from_protobuf_file(const std::string& path);
-
-// --- Human-readable text format (debug only) ---
-
-std::string export_to_protobuf_text(const KvChunkAddressTable& table);
-void export_to_protobuf_text_file(const KvChunkAddressTable& table, const std::string& path);
-KvChunkAddressTable import_from_protobuf_text(const std::string& text);
-KvChunkAddressTable import_from_protobuf_text_file(const std::string& path);
+using tt::tt_metal::internal::disaggregation::export_to_protobuf_text;
+using tt::tt_metal::internal::disaggregation::export_to_protobuf_text_file;
+using tt::tt_metal::internal::disaggregation::import_from_protobuf_text;
+using tt::tt_metal::internal::disaggregation::import_from_protobuf_text_file;
 
 }  // namespace tt::tt_metal::experimental::disaggregation
