+++ tt_metal/impl/internal/disaggregation/kv_chunk_address_table_protobuf.cpp
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include "tt_metal/impl/experimental/disaggregation/kv_chunk_address_table_protobuf.hpp"
+#include "tt_metal/impl/internal/disaggregation/kv_chunk_address_table_protobuf.hpp"
 
 #include <fstream>
 #include <stdexcept>
@@ -12,7 +12,7 @@
 
 #include "protobuf/kv_chunk_address_table.pb.h"
 
-namespace tt::tt_metal::experimental::disaggregation {
+namespace tt::tt_metal::internal::disaggregation {
 
 namespace detail {
 
@@ -185,4 +185,4 @@ KvChunkAddressTable import_from_protobuf_text_file(const std::string& path) {
     return import_from_protobuf_text(text);
 }
 
-}  // namespace tt::tt_metal::experimental::disaggregation
+}  // namespace tt::tt_metal::internal::disaggregation
